requirements.txt
fastapi==0.116.0
uvicorn==0.34.0
httpx==0.28.1
langchain-core==0.3.68
langchain-community==0.3.27
langchain-openai==0.3.27
langchain-anthropic==0.3.17
langsmith==0.4.4
langgraph==0.5.1
langgraph-checkpoint-sqlite==2.0.6
openai
anthropic
supabase==2.16.0
markdown==3.8.2
python-dotenv==1.1.1
pydantic==2.11.7
pydantic-settings==2.8.1
sentry-sdk[fastapi]==2.20.0
numpy
